# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Service Provider & Consumer API",
    "version": "1.0.0",
    "description": "API for managing service providers and consumers (SQLite version)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
...
  "paths": {
    "/api/sp-create": {
      "post": {
        "summary": "Create a new service provider",
        "tags": ["Service Providers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceProvider"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service provider created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/sp-get/{id}": {
      "get": {
        "summary": "Get a service provider by ID",
        "tags": ["Service Providers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service provider details"
          },
          "404": {
            "description": "Service provider not found"
          }
        }
      }
    },
    "/api/sp-update/{id}": {
      "put": {
        "summary": "Update a service provider by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceProvider"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service provider updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Service provider not found"
          }
        }
      }
    },
    "/api/sp-delete/{id}": {
      "delete": {
        "summary": "Delete a service provider by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service provider deleted successfully"
          },
          "404": {
            "description": "Service provider not found"
          }
        }
      }
    },
    "/api/sp-filter": {
      "post": {
        "summary": "Filter service providers",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceProviderFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filtered list of service providers"
          }
        }
      }
    },
...
    "/api/sp-subscription-status/{id}": {
      "get": {
        "summary": "Get subscription status for a service provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription status"
          },
          "404": {
            "description": "Service provider not found"
          }
        }
      }
    },
    "/api/sp-renew-subscription/{id}": {
      "post": {
        "summary": "Renew service provider subscription",
        "description": "Renew subscription for a specific service provider",
        "tags": ["Service Providers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Service provider ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "months": {
                    "type": "integer",
                    "description": "Number of months to renew",
                    "minimum": 1,
                    "example": 1
                  }
                },
                "required": ["months"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription renewed successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Service provider not found"
          }
        }
      }
    },
...
    "/api/payment-upload": {
      "post": {
        "summary": "Upload payment proof image",
        "tags": ["Service Providers"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment proof image uploaded successfully"
          },
          "400": {
            "description": "Bad request - missing or invalid file"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/images/{filename}": {
      "get": {
        "summary": "Serve payment proof images",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image filename"
          }
        ],
        "responses": {
          "200": {
            "description": "Image served successfully"
          },
          "400": {
            "description": "Filename required"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ServiceProvider": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "city": { "type": "string" },
          "skillset": { "type": "string" },
          "contactNo": { "type": "string" },
          "description": { "type": "string" },
          "experience": { "type": "string" },
          "screenshot": { "type": "string", "description": "Image URL for payment proof" },
          "pin": { 
            "type": "string", 
            "description": "6-digit PIN code",
            "pattern": "^[0-9]{6}$",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": ["name", "city", "skillset", "contactNo", "pin"]
      },
      "ServiceProviderFilter": {
        "type": "object",
        "properties": {
          "city": { "type": "string" },
          "skillset": { "type": "string" },
          "experience": { "type": "string" },
          "name": { "type": "string" },
          "search": { "type": "string" }
        }
      },
...
    }
  }
}